+++ CHANGELOG.md
@@ -1,3 +1,73 @@
+# [3.9.0](https://github.com/matter-labs/zksync-wallet-vue/compare/3.8.1...3.9.0) (2022-03-14)
+
+
+### Features
+
+* redirect to testnet etherscan if used etherscan hypelink ([#230](https://github.com/matter-labs/zksync-wallet-vue/issues/230)) ([cf1aced](https://github.com/matter-labs/zksync-wallet-vue/commit/cf1aced051cd668433427c024faedf263a1204c1))
+
+## [3.8.1](https://github.com/matter-labs/zksync-wallet-vue/compare/3.8.0...3.8.1) (2022-02-28)
+
+
+### Bug Fixes
+
+* expect CIDv0 for NFTs ([#229](https://github.com/matter-labs/zksync-wallet-vue/issues/229)) ([d4f601b](https://github.com/matter-labs/zksync-wallet-vue/commit/d4f601b7fb6975488874bc24b6fefe8329cbd986))
+
+# [3.8.0](https://github.com/matter-labs/zksync-wallet-vue/compare/3.7.0...3.8.0) (2022-02-09)
+
+
+### Features
+
+* deposit with orbiter ([#224](https://github.com/matter-labs/zksync-wallet-vue/issues/224)) ([b769466](https://github.com/matter-labs/zksync-wallet-vue/commit/b769466dc954e189171d00e9ce39edbbb3488553))
+
+# [3.7.0](https://github.com/matter-labs/zksync-wallet-vue/compare/3.6.0...3.7.0) (2022-02-09)
+
+
+### Features
+
+* sentry tunnel lambda function ([#220](https://github.com/matter-labs/zksync-wallet-vue/issues/220)) ([1ac8cdd](https://github.com/matter-labs/zksync-wallet-vue/commit/1ac8cdd34946dce370f6eb2c73e9983a07a9a468))
+
+# [3.6.0](https://github.com/matter-labs/zksync-wallet-vue/compare/3.5.0...3.6.0) (2022-02-04)
+
+
+### Features
+
+* mixpanel proxy ([#223](https://github.com/matter-labs/zksync-wallet-vue/issues/223)) ([9962ef1](https://github.com/matter-labs/zksync-wallet-vue/commit/9962ef15f5b3dc653c384bf702d34eb4fa5a2ce3))
+
+# [3.5.0](https://github.com/matter-labs/zksync-wallet-vue/compare/3.4.3...3.5.0) (2022-02-01)
+
+
+### Features
+
+* banxa & moonpay functions added ([#213](https://github.com/matter-labs/zksync-wallet-vue/issues/213)) ([720b68c](https://github.com/matter-labs/zksync-wallet-vue/commit/720b68c4082e2059acbabf39d58bec6611603b74))
+
+## [3.4.3](https://github.com/matter-labs/zksync-wallet-vue/compare/3.4.2...3.4.3) (2022-01-27)
+
+
+### Bug Fixes
+
+* remove gitter ([#215](https://github.com/matter-labs/zksync-wallet-vue/issues/215)) ([3f0e99b](https://github.com/matter-labs/zksync-wallet-vue/commit/3f0e99b9d64e2f565d3810ba9d672b064a96b756))
+
+## [3.4.2](https://github.com/matter-labs/zksync-wallet-vue/compare/3.4.1...3.4.2) (2022-01-25)
+
+
+### Bug Fixes
+
+* click on nft image not redirecting to nft page ([#211](https://github.com/matter-labs/zksync-wallet-vue/issues/211)) ([4e97b17](https://github.com/matter-labs/zksync-wallet-vue/commit/4e97b174e14d974b3031d14aa69af20746e2741d))
+
+## [3.4.1](https://github.com/matter-labs/zksync-wallet-vue/compare/3.4.0...3.4.1) (2022-01-25)
+
+
+### Bug Fixes
+
+* docs links ([#212](https://github.com/matter-labs/zksync-wallet-vue/issues/212)) ([e9337a3](https://github.com/matter-labs/zksync-wallet-vue/commit/e9337a303470545fcfd3e61ecbc3d2b9d94c64d0))
+
+# [3.4.0](https://github.com/matter-labs/zksync-wallet-vue/compare/3.3.0...3.4.0) (2022-01-20)
+
+
+### Features
+
+* nft image previews and use of metadata ([#203](https://github.com/matter-labs/zksync-wallet-vue/issues/203)) ([5d5af05](https://github.com/matter-labs/zksync-wallet-vue/commit/5d5af052830995aa5003a2f1106bcced692b7208))
+
 # [3.3.0](https://github.com/matter-labs/zksync-wallet-vue/compare/3.2.1...3.3.0) (2022-01-18)